+++ tests/integration/api/test_xai.py
@@ -47,11 +47,12 @@ def test_forward_explain(
     is_semisl = model_name.endswith("_semisl")
     task = recipe_split[-2] if not is_semisl else recipe_split[-3]
 
+    if "maskdino" in model_name:
+        # TODO(Eugene): maskdino not support yet.
+        pytest.skip(f"There's issue with inst-seg: {model_name}. Skip for now.")
+
     if "dino" in model_name:
         pytest.skip("DINO is not supported.")
-
-    if "maskrcnn_r50_tv" in model_name:
-        pytest.skip("MaskRCNN R50 Torchvision model doesn't support explain.")
 
     if "rtmdet_tiny" in recipe:
         # TODO (sungchul): enable xai for rtmdet_tiny (CVS-142651)
@@ -113,9 +114,9 @@ def test_predict_with_explain(
     if "dino" in model_name:
         pytest.skip("DINO is not supported.")
 
-    if "instance_segmentation" in recipe:
-        # TODO(Eugene): figure out why instance segmentation model fails after decoupling.
-        pytest.skip("There's issue with instance segmentation model. Skip for now.")
+    if "maskdino" in model_name:
+        # TODO(Eugene): maskdino not support yet.
+        pytest.skip(f"There's issue with inst-seg: {model_name}. Skip for now.")
 
     if "rtmdet_tiny" in recipe:
         # TODO (sungchul): enable xai for rtmdet_tiny (CVS-142651)