+++ CHANGELOG.md
@@ -15,6 +15,8 @@ All notable changes to this project will be documented in this file.
   (https://github.com/openvinotoolkit/training_extensions/pull/3993)
 - Update to work torch compile in detection
   (https://github.com/openvinotoolkit/training_extensions/pull/4003)
+- Fix MaskRCNN/RTMDet-Inst/MaskRCNNTV Explain Mode
+  (https://github.com/openvinotoolkit/training_extensions/pull/4053)
 
 ## \[2.3.0\]